Extract owner-permission logic into its own module for direct unit testing
Moves the permission-normalization logic out of extract.js into:
- lib/ensure-owner-permissions.js - the pure `(entry, opts)` normalizer
- lib/entry-mode.js - shared getEntryMode() + POSIX stat constants

This lets the permission tests assert on the function directly with synthetic
entries instead of patching Module._load to mock extract-zip just to drive the
onEntry callback. The four affected tests are now plain, synchronous unit tests
with no module mocking, and the Windows directory-attribute path (previously
impractical to reach through the extract() pipeline) is now covered too.

Behavior is unchanged; the real-archiver extraction tests still cover the
end-to-end path where extract-zip honors the rewritten attributes.
