+++ packages/zip/lib/extract.js
@@ -1,103 +1,8 @@
 const errors = require('@tryghost/errors');
+const { S_IFMT, S_IFLNK, getEntryMode } = require('./entry-mode');
+const ensureOwnerPermissions = require('./ensure-owner-permissions');
 
 const defaultOptions = {};
-
-// POSIX `stat` mode constants, matching the ones extract-zip uses internally
-// when it derives an entry's type and permissions from its external attributes.
-const S_IFMT = 0o170000; // bit mask for the file type bit fields
-const S_IFDIR = 0o040000; // directory
-const S_IFLNK = 0o120000; // symbolic link
-
-// Owner permission bits we guarantee when `ensureOwnerPermissions` is enabled.
-const OWNER_DIR_BITS = 0o700; // rwx, so the tree can be traversed, written and removed
-const OWNER_FILE_BITS = 0o600; // rw, so files can be read and rewritten/removed
-
-// Reads the POSIX mode an entry carries in the high 16 bits of its external
-// file attributes, exactly as extract-zip does when extracting.
-function getEntryMode(entry) {
-    return (entry.externalFileAttributes >> 16) & 0xffff;
-}
-
-// Mirrors extract-zip's directory detection: POSIX directory type bit, a
-// trailing slash, or the Windows directory attribute.
-function isDirectoryEntry(entry, mode) {
-    if ((mode & S_IFMT) === S_IFDIR) {
-        return true;
-    }
-
-    if (entry.fileName.endsWith('/')) {
-        return true;
-    }
-
-    // Windows' way of specifying a directory
-    // https://github.com/maxogden/extract-zip/issues/13#issuecomment-154494566
-    const madeBy = entry.versionMadeBy >> 8;
-    return madeBy === 0 && entry.externalFileAttributes === 16;
-}
-
-// Resolves the mode extract-zip would synthesise for an entry that carries no
-// POSIX mode bits, mirroring its getExtractedMode() fallback: the caller's
-// defaultDirMode/defaultFileMode, or 0o755/0o644 when those are not set.
-function resolveDefaultMode(isDir, opts) {
-    let mode = 0;
-
-    if (isDir) {
-        if (opts.defaultDirMode) {
-            mode = parseInt(opts.defaultDirMode, 10);
-        }
-
-        if (!mode) {
-            mode = 0o755;
-        }
-    } else {
-        if (opts.defaultFileMode) {
-            mode = parseInt(opts.defaultFileMode, 10);
-        }
-
-        if (!mode) {
-            mode = 0o644;
-        }
-    }
-
-    return mode;
-}
-
-// Normalises an entry's permissions in place so the extracting user can always
-// read, move and remove the result. Only the in-memory entry handed to
-// extract-zip is changed; the original zip file is never rewritten.
-function ensureOwnerPermissions(entry, opts) {
-    let mode = getEntryMode(entry);
-    const isDir = isDirectoryEntry(entry, mode);
-
-    // No POSIX mode bits are present, so extract-zip would synthesise the mode
-    // from defaultDirMode/defaultFileMode (falling back to 0o755/0o644). Resolve
-    // that same default here so the owner bits below are guaranteed even when the
-    // caller-supplied defaults omit them.
-    if (mode === 0) {
-        mode = resolveDefaultMode(isDir, opts);
-    }
-
-    const ownerBits = isDir ? OWNER_DIR_BITS : OWNER_FILE_BITS;
-    let nextMode = mode | ownerBits;
-
-    // A directory that was only inferred (trailing slash / Windows attribute)
-    // may lack the POSIX directory type bit. Set it so extract-zip still
-    // classifies the rewritten mode as a directory.
-    if (isDir && (nextMode & S_IFMT) !== S_IFDIR) {
-        nextMode = (nextMode & ~S_IFMT) | S_IFDIR;
-    }
-
-    // The mode extract-zip would already produce grants the owner the required
-    // access and needs no type-bit fix, so leave the entry untouched.
-    if (nextMode === mode) {
-        return;
-    }
-
-    // Write the new POSIX mode back into the high 16 bits while preserving the
-    // low 16 bits (DOS attributes and other zip-specific flags).
-    const lowBits = entry.externalFileAttributes & 0xffff;
-    entry.externalFileAttributes = (((nextMode & 0xffff) << 16) | lowBits) >>> 0;
-}
 
 function normalizeLimit(value, fieldName) {
     if (value === undefined || value === null || value === Infinity) {
